enforce one-class-per-file convention
